+++ aiohttp/client.py
@@ -302,6 +302,7 @@ class ClientSession:
                    receive_timeout=None,
                    autoclose=True,
                    autoping=True,
+                   heartbeat=None,
                    auth=None,
                    origin=None,
                    headers=None,
@@ -315,6 +316,7 @@ class ClientSession:
                              receive_timeout=receive_timeout,
                              autoclose=autoclose,
                              autoping=autoping,
+                             heartbeat=heartbeat,
                              auth=auth,
                              origin=origin,
                              headers=headers,
@@ -328,6 +330,7 @@ class ClientSession:
                     receive_timeout=None,
                     autoclose=True,
                     autoping=True,
+                    heartbeat=None,
                     auth=None,
                     origin=None,
                     headers=None,
@@ -420,7 +423,8 @@ class ClientSession:
                                            autoping,
                                            self._loop,
                                            time_service=self.time_service,
-                                           receive_timeout=receive_timeout)
+                                           receive_timeout=receive_timeout,
+                                           heartbeat=heartbeat)
 
     def _prepare_headers(self, headers):
         """ Add default headers and transform it to CIMultiDict
@@ -612,10 +616,12 @@ class _RequestContextManager(_BaseRequestContextManager):
     if PY_35:
         @asyncio.coroutine
         def __aexit__(self, exc_type, exc, tb):
-            if exc_type is not None:
-                self._resp.close()
-            else:
-                yield from self._resp.release()
+            # We're basing behavior on the exception as it can be caused by
+            # user code unrelated to the status of the connection.  If you
+            # would like to close a connection you must do that
+            # explicitly.  Otherwise connection error handling should kick in
+            # and close/recycle the connection as required.
+            yield from self._resp.release()
 
 
 class _WSRequestContextManager(_BaseRequestContextManager):