+++ aiohttp/cookiejar.py
@@ -4,13 +4,12 @@ import pickle
 import re
 from collections import defaultdict
 from collections.abc import Mapping
-from http.cookies import Morsel, SimpleCookie
+from http.cookies import Morsel
 from math import ceil
-
 from yarl import URL
 
 from .abc import AbstractCookieJar
-from .helpers import is_ip_address
+from .helpers import SimpleCookie, is_ip_address
 
 
 class CookieJar(AbstractCookieJar):