+++ aiohttp/test_utils.py
@@ -18,7 +18,7 @@ import aiohttp
 from aiohttp.client import _RequestContextManager
 
 from . import ClientSession, hdrs
-from .helpers import sentinel
+from .helpers import TimeService, sentinel
 from .protocol import HttpVersion, RawRequestMessage
 from .signals import Signal
 from .web import Application, Request, Server, UrlMappingMatchInfo
@@ -201,6 +201,7 @@ class TestClient:
         if cookie_jar is None:
             cookie_jar = aiohttp.CookieJar(unsafe=True,
                                            loop=self._loop)
+        kwargs['time_service'] = TimeService(self._loop, interval=0.1)
         self._session = ClientSession(loop=self._loop,
                                       cookie_jar=cookie_jar,
                                       **kwargs)
@@ -384,14 +385,21 @@ class AioHTTPTestCase(unittest.TestCase):
 
     def setUp(self):
         self.loop = setup_test_loop()
+
         self.app = self.loop.run_until_complete(
             self.get_application(self.loop))
-        self.client = TestClient(self.app)
+        self.client = self.loop.run_until_complete(self._get_client(self.app))
+
         self.loop.run_until_complete(self.client.start_server())
 
     def tearDown(self):
         self.loop.run_until_complete(self.client.close())
         teardown_test_loop(self.loop)
+
+    @asyncio.coroutine
+    def _get_client(self, app):
+        """Return a TestClient instance."""
+        return TestClient(self.app)
 
 
 def unittest_run_loop(func):