+++ CHANGES.rst
@@ -4,6 +4,11 @@ CHANGES
 1.3.0 (XXXX-XX-XX)
 ------------------
 
+- Multipart writer validates the data on append instead of on a request send #920
+
+- Multipart reader accepts multipart messages with or without their epilogue
+  to consistently handle valid and legacy behaviors #1526 #1581
+
 - Separate read + connect + request timeouts # 1523
 
 - Fix polls demo run application #1487
@@ -15,8 +20,10 @@ CHANGES
 - Do not use readline when reading the content of a part
   in the multipart reader #1535
 
-- Added `receive_timeout` timeout for websocket to receive complete message. #1024 #1325
-  
+- Added `receive_timeout` timeout for websocket to receive complete message. #1325
+
+- Added `heartbeat` parameter for websocket to automatically send `ping` message. #1024 #777
+
 - Remove `web.Application` dependency from `web.UrlDispatcher` #1510
 
 - Accepting back-pressure from slow websocket clients #1367
@@ -30,14 +37,26 @@ CHANGES
 - Allow users to specify what should happen to decoding errors
   when calling a responses `text()` method #1542
 
+- Back port std module `http.cookies` for python3.4.2 #1566
+
+- Maintain url's fragment in client response #1314
+
+- Allow concurrently close WebSocket connection #754
+
+- Gzipped responses with empty body raises ContentEncodingError #609
+
 - Return 504 if request handle raises TimeoutError.
 
 - Refactor how we use keep-alive and clone lingering timeouts.
 
+- Close response connection if we can not consume whole http
+  message during client response release
+
+- Log warning instead of RuntimeError is websocket connection is closed.
+  
 - Fix ClientSession is not aware of TestClient #1499
 
 - Fixed warnings showing when giving ClientSession an event loop when called from a normal function. #[1468](https://github.com/KeepSafe/aiohttp/pull/1468#issuecomment-269112177)
-
 
 1.2.0 (2016-12-17)
 ------------------